+++ l10n/af.js
@@ -5,9 +5,13 @@ OC.L10N.register(
     "Cancel" : "Cancel",
     "Update" : "Werk by",
     "Comments" : "Kommentare",
+    "Dismiss" : "Ontslaan",
     "Save" : "Stoor",
     "Add a comment" : "Voeg kommentaar toe",
     "Send" : "Stuur",
+    "Unknown user" : "Onbekende gebruiker",
+    "File" : "Lêer",
+    "min" : "min",
     "Start" : "Begin",
     "Reset" : "Herstel",
     "Open" : "Open"