+++ l10n/bn_BD.js
@@ -1,6 +1,7 @@
 OC.L10N.register(
     "whiteboard",
     {
+    "Close" : "বন্ধ",
     "Cancel" : "Cancel",
     "Update" : "পরিবর্ধন",
     "Edit" : "সম্পাদনা",
@@ -8,6 +9,9 @@ OC.L10N.register(
     "Add a comment" : "মন্তব্য দিন",
     "Send" : "পাঠাও",
     "Unknown" : "অজানা",
+    "Unknown user" : "অপরিচিত ব্যবহারকারী",
+    "Hours" : "ঘন্টা",
+    "Minutes" : "মিনিট",
     "Reset" : "পূণঃনির্ধানণ",
     "Pause" : "বিরতি"
 },