+++ l10n/bs.js
@@ -1,6 +1,7 @@
 OC.L10N.register(
     "whiteboard",
     {
+    "Close" : "Zatvori",
     "Cancel" : "Cancel",
     "Update" : "Ažuriraj",
     "Comments" : "Komentari",
@@ -9,6 +10,8 @@ OC.L10N.register(
     "Add a comment" : "Dodaj komentar",
     "Send" : "Pošalji",
     "Unknown" : "Nepoznato",
+    "Hours" : "Sati",
+    "Minutes" : "Minute",
     "Pause" : "Pauza"
 },
 "nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2);");