+++ l10n/cy_GB.js
@@ -2,6 +2,7 @@ OC.L10N.register(
     "whiteboard",
     {
     "Whiteboard" : "Bwrdd gwyn",
+    "Close" : "Cau",
     "Cancel" : "Cancel",
     "Update" : "Diweddaru",
     "Voting" : "Pleidleisio",
@@ -10,9 +11,12 @@ OC.L10N.register(
     "Send" : "Anfon",
     "Connecting" : "Yn cysylltu",
     "Unknown" : "Anhysbys",
+    "Duration" : "Hyd",
     "Running" : "Rhedeg",
+    "min" : "mun",
     "Pause" : "Seibio",
     "Status" : "Statws",
-    "Open" : "Ar Agor"
+    "Open" : "Ar Agor",
+    "Vote" : "Pleidleisio"
 },
 "nplurals=4; plural=(n==1) ? 0 : (n==2) ? 1 : (n != 8 && n != 11) ? 2 : 3;");