+++ l10n/nn_NO.js
@@ -2,14 +2,19 @@ OC.L10N.register(
     "whiteboard",
     {
     "Advanced settings" : "Avanserte innstillingar",
+    "Close" : "Lat att",
     "Cancel" : "Cancel",
     "Update" : "Oppdater",
     "Comments" : "Kommentarar",
+    "Dismiss" : "Forkast",
+    "Guest" : "Gjest",
     "Edit" : "Rediger",
     "Save" : "Lagre",
     "Reply" : "Svare",
     "Send" : "Send",
     "Unknown" : "Ukjend",
+    "Unknown user" : "Ukjend brukar",
+    "File" : "Fil",
     "Pause" : "Pause"
 },
 "nplurals=2; plural=(n != 1);");